pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Replit user <replituser@example.com>"]
description = ""
name = "spotifytoyoutubebot"
version = "0.1.0"
[tool.poetry.dependencies]
discord = "^1.0.1"
python = "^3.8"
spotipy = "^2.17.1"
youtube-python = "^1.0.13"
flask = "^1.1.2"
[tool.poetry.dev-dependencies]
